requirements.txt
torch>=2.0
torchvision>=0.15
scikit-learn>=1.0
opencv-python>=4.5
pandas>=1.3
numpy>=1.21
tqdm>=4.60
matplotlib>=3.4
PyYAML>=6.0
wandb>=0.15
albumentations>=1.3
seaborn>=0.11
pytest>=7.0
PyQt5>=5.15
mlflow>=2.0
Pillow>=9.0
pydicom>=2.4
